Feat(Service): add margin error in collector sampling

+++ src/main/java/com/cms/audit/api/Sampling/dto/request/CollectorSamplingDto.java
@@ -23,6 +23,6 @@ public class CollectorSamplingDto {
 
     private Long unit_sampling_unit;
 
-    private Long margin_error;
+    private Double margin_error;
 
 }
